junit: Fix parsing of test names
The test names parsing logic here was a bit loose previously with light
testing. As a result it could exclude valid test names in the form of a
filepath (for example, including a period character), and if used it
would always return only two names matching the first set of
match+submatch for the regex.

Fix it, make the code more robust to arbitrary input, and write some
basic tests to validate the functionality. Upcoming commits will rely on
this being accurate.

Signed-off-by: Joe Stringer <joe@isovalent.com>